global-tools/requirements.txt
fastapi
uvicorn[standard]
gunicorn
tavily-python
pymongo
email-validator
chromadb
numpy
python-dotenv
requests